// web-old/src/types.ts
export interface RoomDetailType {
    RoomNo: number;
    RoomType: string;
    RoomTypeCode: number;
    RoomTypeSupplierCode: string;
    MealPlanSupplierCode: string;
    RoomStatus: string;
    BlackOut: {
        Status: string;
        Msg: string;
    };
    CurrCode: string;
    ContractTokenId: string;
    RoomConfigurationId: number;
    RatePlanCode: string;
    RatePlanId: number;
    MealPlan: string;
    MealPlanCode: string;
    NumberOfMeals: number;
    RoomNumber: number;
    Rate: number;
    RoomStatusDetails: {
        Status: string[];
    };
    DiscountDetails: {};
    CancellationPolicyDetails: {
        // time is in local time of the location
        Cancellation: {
            FromDate: number;
            FromTime: string;
            ToDate: number;
            NightToCharge: number;
        }[];
    };
    PromotionalContract: string;
    PackageYN: string;
    NonRefundable: string;
    DynamicYN: string;
    TotalRate: number;
    RateBeforeTax: number;
    TotalDiscount: number;
    RecommendedRetailPrice: number;
    ContractLabel: string;
    MealPlanCodeLong: number;
}

export interface HotelSearchItemType {
    SourceId: number;
    HotelId: number;
    HotelName: string;
    PreferredStatus: string;
    PropertyType: string;
    StarRating: number;
    GeoLocation: {
        Longitude: string;
        Latitude: string;
    };
    Chain: string;
    HotelCode: string;
    TimeZone: string;
    City: string;
    RoomTypeDetails: {
        Rooms: {
            Room: RoomDetailType[];
        };
    };
    StartDate: number;
    EndDate: number;
    Country: string;
    Restriction: string;
}
// ...
export interface RoomCfgType {
    rooms: {
        // FIX: why are adults and children passed differently...
        adults: number;
        children: {
            age: number;
        }[];
    }[];
}